Stop returning an error from IndexCounter and IndexCounterWithAttempts
Returning the error is problematic in our case since all usages have the form:

  Eventually(IndexCounter(...)).Should(Equal(...))

Which usually fails with cryptic error message if the second return
value (i.e. the error) isn't nil.

  Error: Unexpected non-nil/non-zero extra argument at index 1:
   <*url.Error>: &url.Error{Op:"Get", URL:"http://host/index", Err:(*net.OpError)(0xabcd)}

This has a caveat. The failure line will be in the helpers instead of the
original test, but I'm hoping this will be a rare case and will be obvious
where the error occurred. We cannot use `ExpectWithOffset' since the
GetIndexFromEndpointFor and IndexCounterWithAttempts can be called with
different stack depth.