wip: add tentative test for readiness health check
This commit adds an acceptance test to see when an LRP is defined with
healthchecks that the healthcheck binary will fail with the correct
error message, indicating that in fact a readiness check is being run.

This will not work until we make other changes that allow the LRP to be
created with the readiness health check, most likely in the `executor`

These changes are part of a [larger app-health check
initiative](https://docs.google.com/document/d/1AS8GzKg3pdAv56nySQTgF4NaqqzrBwmztkL2_8ahKa4/edit#heading=h.jmuxsjek21yu).

[#185035396](https://www.pivotaltracker.com/story/show/185035396)

Signed-off-by: Josh Russett <jrussett@vmware.com>